Move FilteredList into widgets/view/filter_list module; update FilterListView
... and it's still broken: views over an EditBox don't pass updates
